Refactor: clean InfoPopover styles
