+++ src/components/shared/TokenInfoPopover/TokenInfoPopover.css.d.ts
@@ -1,14 +1,11 @@
-declare namespace InfoPopoverCssNamespace {
-  export interface IInfoPopoverCss {
+declare namespace TokenInfoPopoverCssNamespace {
+  export interface ITokenInfoPopoverCss {
     addToWallet: string;
     address: string;
-    badges: string;
-    container: string;
     displayName: string;
     etherscanLink: string;
     main: string;
     nativeTokenMessage: string;
-    reputationHeading: string;
     section: string;
     symbol: string;
     textContainer: string;
@@ -16,9 +13,9 @@ declare namespace InfoPopoverCssNamespace {
   }
 }
 
-declare const InfoPopoverCssModule: InfoPopoverCssNamespace.IInfoPopoverCss & {
+declare const TokenInfoPopoverCssModule: TokenInfoPopoverCssNamespace.ITokenInfoPopoverCss & {
   /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
-  locals: InfoPopoverCssNamespace.IInfoPopoverCss;
+  locals: TokenInfoPopoverCssNamespace.ITokenInfoPopoverCss;
 };
 
-export = InfoPopoverCssModule;
+export = TokenInfoPopoverCssModule;
